Account Import CLI refactor
